@@ -8,7 +8,7 @@ publish_to: 'none'
 version: 1.0.0+3
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -16,27 +16,26 @@ dependencies:
 
 
 
-  cupertino_icons: ^0.1.3
-  dio: ^3.0.10
-  adaptive_theme: ^1.0.0
-  share: ^0.6.5
-  shared_preferences: ^0.5.10
-  provider: ^4.3.2+2
-  get: ^3.10.2
-  sqflite: ^1.3.1+1
-  expandable: ^4.1.4
-  path_provider: ^1.6.14
-  flutter_widget_from_html: ^0.5.0+3
-  hive: ^1.4.4
-  hive_flutter: ^0.3.1
-  http: ^0.12.2
-  in_app_update: ^1.1.13
+  cupertino_icons: ^1.0.2
+  adaptive_theme: ^2.3.0
+  share: ^2.0.4
+  shared_preferences: ^2.0.6
+  provider: ^6.0.0
+  get: ^4.3.8
+  sqflite: ^2.0.0+3
+  expandable: ^5.0.1
+  path_provider: ^2.0.0
+  flutter_html: ^2.1.5
+  hive: ^2.0.4
+  hive_flutter: ^1.1.0
+  http: ^0.13.0
+
 
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^0.7.1
+  hive_generator: ^1.1.1
   build_runner:
 
 # dependency_overrides:
