@@ -6,11 +6,11 @@ import 'package:stg_app/models/SubItem.dart';
 class HTMLDataModel extends ChangeNotifier {
   HTMLDataModel();
 
-  Future<HTMLItem> loadHTML(SubItem subItem) async {
+  Future<HTMLItem?> loadHTML(SubItem subItem) async {
     return loadDatabaseItem(subItem);
   }
 
-  Future<HTMLItem> loadDatabaseItem(SubItem subItem) async {
+  Future<HTMLItem?> loadDatabaseItem(SubItem subItem) async {
     final box = Hive.lazyBox<HTMLItem>('htmlItemBox');
     return box.get(subItem.address);
   }
